+++ lib/flutter_kanpur_ui_kit.dart
@@ -2,6 +2,8 @@
 ///
 /// This package provides customizable UI widgets such as
 /// bottom navigation bars and shared design utilities.
+/// gradient button which gives you 3D feel.
 library flutter_kanpur_ui_kit;
 
 export 'src/bottom_nav_bar.dart';
+export 'src/gradient_button.dart';